+++ src/components/overlays/InviteForm/index.tsx
@@ -34,6 +34,7 @@ import debounce from 'lodash.debounce'
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import Patterns from 'types/Patterns'
+import { ProcessingType } from 'components/pages/InviteBusinessPartner'
 
 interface AddInviteFormOverlayProps {
   openDialog?: boolean
@@ -141,7 +142,7 @@ export const InviteForm = ({
           >
             {`${t('global.actions.cancel')}`}
           </Button>
-          {state === 'busy' ? (
+          {state === ProcessingType.BUSY ? (
             <LoadingButton
               color="primary"
               helperText=""