+++ src/components/pages/Admin/components/RegistrationRequests/index.tsx
@@ -140,7 +140,7 @@ export default function RegistrationRequests() {
 
   const onUpdateBpn = async (bpn: string) => {
     setIsLoading(true)
-    await updateBpn({ bpn: bpn, applicationId: selectedRequestId })
+    await updateBpn({ bpn, applicationId: selectedRequestId })
       .unwrap()
       .then((payload) => {
         setEnableBpnInput(false)
@@ -181,10 +181,8 @@ export default function RegistrationRequests() {
         <CompanyDetailOverlay
           {...{
             openDialog: overlayOpen,
-            selectedRequestId: selectedRequestId,
-            handleOverlayClose: () => {
-              setOverlayOpen(false)
-            },
+            selectedRequestId,
+            handleOverlayClose: () => { setOverlayOpen(false) },
           }}
         />
       )}