+++ src/components/pages/InviteBusinessPartner/components/InviteList/index.tsx
@@ -33,6 +33,10 @@ import dayjs from 'dayjs'
 import { setSearchInput } from 'features/appManagement/actions'
 import { updateInviteSelector } from 'features/control/updates'
 
+interface FetchHookArgsType {
+  expr: string
+}
+
 export const InviteList = ({
   fetchHook,
   fetchHookArgs,
@@ -40,7 +44,7 @@ export const InviteList = ({
   searchExpr,
 }: {
   fetchHook: (paginArgs: PaginFetchArgs) => any
-  fetchHookArgs?: any
+  fetchHookArgs?: FetchHookArgsType
   onSearch?: (search: string) => void
   searchExpr?: string
 }) => {
@@ -57,7 +61,7 @@ export const InviteList = ({
 
   return (
     <section id="identity-management-id">
-      <PageLoadingTable<CompanyInvite>
+      <PageLoadingTable<CompanyInvite, FetchHookArgsType>
         searchExpr={searchExpr}
         toolbarVariant="premium"
         searchPlaceholder={t('global.table.searchName')}