+++ src/components/pages/NotificationCenter/index.tsx
@@ -147,7 +147,7 @@ export default function NotificationCenter() {
       setLoaded(false)
     }
     setNotificationState({
-      page: page,
+      page,
       size: PAGE_SIZE,
       args: {
         notificationTopic: filterOption,