+++ CHANGELOG.md
@@ -13,7 +13,7 @@
   - Enable search & filter
 - Technical User Details
   - Enable credential reset
-- Serview Overview
+- Service Overview
   - Added Sub menu for active services in service overview
   - Added Deactivate functionality
   - Added image to service overview cards
@@ -31,8 +31,6 @@
   - Last section background color issue
 - My Account
   - Show success message on click of copy to clip board
-- Invite Business Partner Form
-  - Fix loader position to center
 - Page Loading and Error Component
   - Show loading component and based on the response show "No items" component or ErrorBar component
 - App management
@@ -51,6 +49,12 @@
 - Service Subscription
   - UI Changes
   - Add filters
+- Enable ESLINT rule
+  - import/no-duplicates
+  - prefer-const
+  - object shorthand
+- Notifications
+  - App Subscription Request content update
 - Digital Twin
   - Code clean up
 - Usermanagement
@@ -59,6 +63,9 @@
   - Change Subscription roles
 - Service release process
   - Clearing fields while registering new service
+  - Skip technical integration for consultancy service
+- Shared
+  - Removed PageLoadingTable component
 - Linter Findings
   - No confusing void expression fixes